zero leaks

+++ libft/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: dinda-si <dinda-si@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/04 12:55:55 by dinda-si          #+#    #+#             */
-/*   Updated: 2023/10/16 15:28:11 by dinda-si         ###   ########.fr       */
+/*   Updated: 2024/11/12 17:31:31 by dinda-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strdup(const char *s)
 {
 	char	*dest;
-	size_t	i;
+	int		i;
 
 	dest = malloc((ft_strlen(s) + 1));
 	i = 0;
